requirements.txt
langchain>=0.0.281
openai==0.28 
tqdm>=4.65.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.3.2
chromadb>=0.3.29
tiktoken>=0.4.0
pytest-asyncio
litellm>=0.1.558
Pydantic==1.10.7
faiss-cpu==1.7.4
pinecone-client==2.2.4
python-dotenv
requests
python-dateutil
pyowm
duckduckgo-search
xmltodict
wikipedia
arxiv
numexpr
uvicorn
fastapi

-e .